Restructure folder structure (#9)
* New folder structure

* Moved a few things around

* Reorganize color spaces into folders

* Remove extra method

* Updated other files

* SwiftFormat

* Simplify comments
