+++ Package.swift
@@ -10,14 +10,14 @@ let package = Package(
         .macOS(.v14),
         .tvOS(.v17),
         .watchOS(.v10),
-        .visionOS(.v1)
+        .visionOS(.v1),
     ],
     products: [
         // Products define the executables and libraries a package produces, making them visible to other packages.
         .library(
             name: "ColorTokensKit",
             targets: ["ColorTokensKit"]
-        )
+        ),
     ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
@@ -25,12 +25,12 @@ let package = Package(
         .target(
             name: "ColorTokensKit",
             resources: [
-                .process("Resources/ColorPalettes.json")
+                .process("Resources/ColorPalettes.json"),
             ]
         ),
         .testTarget(
             name: "ColorTokensKitTests",
             dependencies: ["ColorTokensKit"]
-        )
+        ),
     ]
 )
