+++ Sources/ColorTokensKit/ColorSpace/XYZ/XYZColor+Manipulation.swift
@@ -1,0 +1,10 @@
+//
+//  XYZColor+Manipulation.swift
+//  ColorTokensKit
+//
+
+import Foundation
+
+public extension XYZColor {
+    // TODO: Implement manipulation methods
+}
