+++ Sources/ColorTokensKit/Platform/AppKit/NSColor+Dynamic.swift
@@ -1,0 +1,15 @@
+//
+//  NSColor+Dynamic.swift
+//  ColorTokensKit
+//
+
+#if canImport(AppKit)
+    import AppKit
+
+    // TODO: Implement full AppKit support
+    /// Source: https://www.jessesquires.com/blog/2023/07/11/creating-dynamic-colors-in-swiftui/
+
+    public extension NSColor {
+        // Placeholder for future AppKit implementation
+    }
+#endif
