+++ Sources/ColorTokensKit/Platform/SwiftUI/Color+Initialization.swift
@@ -1,0 +1,40 @@
+//
+//  Color+Initialization.swift
+//  ColorTokensKit
+//
+
+import SwiftUI
+
+public extension Color {
+    /// Initialize from HSL values
+    init(h hue: Double, s saturation: Double, l lightness: Double, a opacity: Double = 1) {
+        let brightness = lightness + saturation * min(lightness, 1 - lightness)
+        let saturation = brightness == 0 ? 0 : 2 * (1 - lightness / brightness)
+        self.init(hue: hue / 360, saturation: saturation, brightness: brightness, opacity: opacity)
+    }
+
+    /// Initialize from hex string
+    init(hex: String) {
+        let hex = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
+        var int: UInt64 = 0
+        Scanner(string: hex).scanHexInt64(&int)
+        let a, r, g, b: UInt64
+        switch hex.count {
+        case 3: // RGB (12-bit)
+            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
+        case 6: // RGB (24-bit)
+            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
+        case 8: // ARGB (32-bit)
+            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
+        default:
+            (a, r, g, b) = (1, 1, 1, 0)
+        }
+        self.init(
+            .sRGB,
+            red: Double(r) / 255,
+            green: Double(g) / 255,
+            blue: Double(b) / 255,
+            opacity: Double(a) / 255
+        )
+    }
+}
