+++ Sources/ColorTokensKit/Platform/SwiftUI/Color+ProColors.swift
@@ -16,8 +16,6 @@ import Foundation
 import SwiftUI
 
 public extension Color {
-    // MARK: - Pro Colors
-
     // Pro color getters
     static var proGray: LCHColor { LCHColor.getPrimaryColor(forHue: 0, isGrayscale: true) }
     static var proPink: LCHColor { LCHColor.getPrimaryColor(forHue: 5) }
@@ -69,7 +67,7 @@ public extension Color {
             "Purple": proPurple,
             "Violet": proViolet,
             "Plum": proPlum,
-            "Ruby": proRuby
+            "Ruby": proRuby,
         ]
     }
 }
