+++ Sources/ColorTokensKit/Services/Palettes/ColorPaletteData.swift
@@ -1,15 +1,5 @@
 import Foundation
 
-struct ColorRamp {
-    let name: String
-    let stops: [String: LCHColor]
-
-    public init(name: String, stops: [String: LCHColor]) {
-        self.name = name
-        self.stops = stops
-    }
-}
-
 struct ColorPaletteData: Codable {
     private var palettes: [String: [String: String]]
